authentication_service/requirements.txt
fastapi
python-decouple
boto3
pymysql
uvicorn
pydantic
pydantic[email]
sqlalchemy
sqlalchemy-utils
cryptography
httpx
python-jose